requirements.txt
Pillow>=10.0.0
openpyxl>=3.1.0
requests>=2.31.0
PyYAML>=6.0
watchdog>=4.0.0
